// frontend/src/features/registry/models/PostRegistryErrorCode.ts
enum PostRegistryErrorCode {
    COLUMN_EMPTY = "COLUMN_EMPTY",
    PHONE_FORMAT_INVALID = "PHONE_FORMAT_INVALID",
    USERNAME_LESS_THAN_2_CHARS = "USERNAME_LESS_THAN_2_CHARS",
    PASSWORD_LESS_THAN_8_CHARS = "PASSWORD_LESS_THAN_8_CHARS",
    USERNAME_CONTAINS_INVALID_CHARS = "USERNAME_CONTAINS_INVALID_CHARS",
    EMAIL_FORMAT_INVALID = "EMAIL_FORMAT_INVALID",
    USERNAME_EXISTS = "USERNAME_EXISTS"
}
export default PostRegistryErrorCode